requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
validators>=0.22.0

# Environment and configuration
python-dotenv>=1.0.0
nest-asyncio>=1.5.8

# Web scraping and content processing
requests>=2.31.0
beautifulsoup4>=4.12.0
unstructured>=0.11.0

# LangChain core and components
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.2

# Document loaders and processors
pypdf>=3.17.0
youtube-transcript-api>=0.6.0
pytube>=15.0.0

# HTTP and async support
httpx>=0.25.0

# Additional utilities
python-multipart>=0.0.6
pydantic>=2.0.0
tenacity>=8.2.0
protobuf==3.20.0

# Optional: Better performance
watchdog>=3.0.0
psutil>=5.9.0
